Methods

The main methods of 'squarebrackets' use the naming convention A_B:
A tells you on what kind of object and what kind of indices the method operates on;
B tells you what operation is performed.

For the A part, the following is available:

i_: operates on subsets of atomic objects by (flat/linear) indices.

i2_: operates on subsets of recursive objects by (flat/linear) indices.

ss_: operates on subsets of atomic objects by (dimensional) subscripts.

ss2_: operates on subsets of recursive objects by (dimensional) subscripts.

slice_: uses index-less, sequence-based, and efficient operations on mutatomic objects.

slicev_: uses index-less, value-based and efficient operations on mutatomic objects.

For the B part, the following is available:

_x: extract, exchange, or duplicate (if applicable) subsets.

_wo: returns the original object without the provided subsets.

_mod: modify subsets and return copy.

_set: modify subsets using pass-by-reference semantics.

_rec: access recursive subsets of lists - ONLY available with the i2_ prefix.

_recin: replace, transform, delete, or extend recursive subsets of lists - ONLY available with the i2_ prefix.

There is also the somewhat separate idx method, which works on both recursive and non-recursive objects.

Thus, for example, the i_x method extracts/exchanges/duplicates subsets of atomic objects by flat/linear indices.

Illustration

To illustrate, let's take the methods used for extracting subsets (\(\ast\)_x):
When y is atomic, the following holds (roughly speaking):

i_x(y, i) corresponds to y[i]

ss_x(y, n(i, k), c(1, 3)) corresponds to y[i, , k]

When y is a list (i.e. recursive), the following holds (roughly speaking):

i2_x(y, i) corresponds to y[i] or y[[i]] (depending on the arguments given in i2_x())

ss2_x(y, n(i, k), c(1, 3)) corresponds to y[i, , k] or y[[i, , k]] (depending on the arguments given in ss2_x())

Finding the Appropriate Help Pages

Now that it is clear what the naming convention is of the main methods, one can easily find out information about a particular method by usign the ? operator.
So to find out about modifying recursive objects by reference using dimensional indices, type in:
?ss2_set

Note that some combinations of A and B are not available.
For example, as the _rec _recin suffices are only applicable for recursive objects using linear indices, only i2_rec and i2_recin exist;
combinations like i_rec, ss_recin, etc. don't exist.
Another example: there is no _mod suffix available for the slice_ and slicev_ methods.