Indexing Fundamentals

This help page explains the fundamentals regarding how 'squarebrackets' treats indexing.

Indexing Types

Base 'R' supports indexing through logical, integer, and character vectors.
'squarebrackets' supports these also (albeit with some improvements), but also supports some additional methods of indexing.

Whole numbers
Whole numbers are the most basic form on index selection.
All forms of indexing in 'squarebrackets' are internally translated to integer (or double if > (2^31 - 1)) indexing first, ensuring consistency.
Indexing through integer/numeric indices in 'squarebrackets' works the same as in base 'R', except that negative values are not allowed.

Logical
Selecting indices with a logical vector in 'squarebrackets' works the same as in base 'R', except that recycling is not allowed.
Thus the logical vector must be of the correct length (i.e. length(x) or dim(x)[L], depending on the situation).

Characters
When selecting indices using a character vector, base 'R' only selects the first matches in the names.
'squarebrackets', however, selects all matches.
Character indices are internally translated to integer indices using match_all.

Imaginary Numbers
A complex vector y is structured as
y = a + b * i
where Re(y) returns a, and Im(y) returns b.
squarebrackets' includes support for indexing through imaginary numbers (Im(y)) of complex vectors.
Indexing with imaginary numbers is a generalization of indexing with regular integers.

It works as follows:
Imaginary numbers that are positive integers, like 1:10 * 1i, work the same as regular integers.
Imaginary numbers that are negative integers, like 1:10 * -1i, index by counting backwards (i.e. from the end), where the integer indices are computed as n + Im(y) + 1L.
Here n is the maximum possible integer (i.e. length(x), or dim(x)[L], depending on the situation), and Im(y) is negative.
Note that only the Imaginary part of a complex vector is used (Im(y));
the Real part (Re(y)) is ignored.

See the results of the following code as an example:


x <- 1:30 # vector of 30 elements

sb_x(x, 1:10 * 1i) # extract first 10 elements
#>  [1]  1  2  3  4  5  6  7  8  9 10

sb_x(x, 1:10 * -1i) # extract last 10 elements
#>  [1] 30 29 28 27 26 25 24 23 22 21

sb_x(x, 10:1 * -1i) # last 10 elements, in tail()-like order
#>  [1] 21 22 23 24 25 26 27 28 29 30

Thus complex vectors allow the user to choose between counting from the beginning, like regular integers, or counting from the end.

Subscripts

One can distinguish between flat indices, array subscripts, and data.frame subscripts.

Flat indices, also called linear indices, specifies the index of a vector, ignoring dimensions (if any are present).
So in an expression like x[i], where i is a vector, i specifies flat indices.

Matrices and arrays also have array subscripts.
Array subscripts works by specifying multiple indexing vectors, which can be of different sizes, where each vector specifies positions in a specific dimension.
Given, for example, a 3-dimensional array, the subscript [1:10, 2:5, 3:9] refers to rows 1 to 10, columns 2 to 5, and layers 3 to 9.
The base S3 vector classes in 'R' use the standard Linear Algebraic convention, as in academic fields like Mathematics and Statistics, in the following sense:

vectors are column vectors (i.e. vertically aligned vectors);

index counting starts at 1;

rows are the first dimension/subscript, columns are the second dimension/subscript, etc.

Thus, the orientation of flat indices in, for example, a 4-rows-by-5-columns matrix, is as follows:

     [,1] [,2] [,3] [,4] [,5]
[1,]    1    5    9   13   17
[2,]    2    6   10   14   18
[3,]    3    7   11   15   19
[4,]    4    8   12   16   20

In a 4 by 4 matrix, subscript [1,2] corresponds to flat index 5.
All array subscripts in 'squarebrackets' also follow this convention.

Data.frame-like objects use data.frame subscripts.
At first glance this may seem the same as the array subscripts of matrices, but they are not:
The column indices of a data.frame-like object is equal to its flat indices.
I.e. for a data.frame, x[i] is essentially the same as x[, i], safe for some attribute handling.
To avoid confusion, 'squarebrackets' does not have an argument for flat indices in its data.frame methods.

Flat indices (or just "indices" for non-dimensional objects) exist for all objects (in data.frame-like objects, flat indices are actually equal to column indices).
Thus flat indices are the "default" indices, and are usually just referred to as simply "indices".

Indexing in Recursive Subsets

One of the differences between atomic and recursive objects, is that recursive objects support recursive subsets, while atomic objects do not.

Bear in mind that every element in a recursive object is a reference to another object.
Consider the following list x:


x <- list(A = letters, B = 1:10, C = list(A = 11:20, B = LETTERS))

Regular subsets, AKA surface-level subset operations ([, [<- in base 'R'), operate on the recursive object itself.
I.e. sb2_x(x, 1), or equivalently x[1], returns the list list(A = letters).
Recursive subset operations ([[, [[<- in base 'R'), on the other hand, operate on an object a subset of the recursive object references to.
I.e. sb2_rec(x, 1), or equivalently x[[1]], returns the character vector letters.

Recursive objects can refer to other recursive objects, which can themselves refer to recursive objects, and so on.
Recursive subsets can go however deep you want.
So, for example, to extract the character vector LETTERS from the aforementioned list x, one would need to do:
sb2_rec(x, c("C","B")), or equivalently, x[["C"]][["B"]].
You can also do this using integers of course: sb2_rec(x, c(3, 2)).

Note that recursive subset operations using sb2_rec/sb2_recin only support positive integer vectors and character vectors;
imaginary numbers (using complex vectors) and logical vectors are not supported.
Moreover, since a recursive subset operation only operates on a single element, specifying the index with a character vector only selects the first matching element (just like base 'R'), not all matches.

Regarding Performance

Integer indices and logical indices are the fastest.
Indexing through names (i.e. character vectors) is the slowest.
Thus if performance is important, use integer or logical indices.