Regarding Modification

This help page describes the main modification semantics available in 'squarebrackets'.

Base R's default modification

For most average users, R's default copy-on-modify semantics are fine.
The benefits of the indexing arguments from 'squarebrackets' can be combined the [<- operator, through the idx method.
The result of the idx() method can be used inside the regular square-brackets operators.
For example like so:

x <- array(...)
my_indices <- idx(x, sub, dims)
x[my_indices] <- value

y <- data.frame(...)
rows <- idx(y, 1:10, 1, inv = TRUE)
cols <- idx(y, c("a", "b"), 2)
y[rows, cols] <- value

thus allowing the user to benefit from the convenient index translations from 'squarebrackets', whilst still using R's default copy-on-modification semantics (instead of the semantics provided by 'squarebrackets').

Explicit Copy

'squarebrackets' provides the sb_mod/sb2_mod method to modify through copy.
This method always copies the modification.
For recursive objects, sb2_mod returns the original object, where only the modified subsets are copied, thus preventing unnecessary usage of memory.

Pass-by-Reference

'squarebrackets' provides the sb_set/sb2_set and slice_set methods to modify by reference, meaning no copy is made at all.
Pass-by-Reference is fastest and the most memory efficient.
But it is also more involved than the other modification forms, and requires more thought.
See squarebrackets_PassByReference for more information.

Arguments rp, tf, .lapply

Using the rp argument in the modification methods, corresponds to something like the following:

x[...] <- rp

Using the tf argument (and .lapply argument, for recursive objects) in the modification methods, corresponds to something like the following:

x[...] <- tf(x[...]) # for atomic objects
x[...] <- .lapply(x[...], tf) # for recursive objects

where tf is a function that returns an object of appropriate type and size (so tf should not be a pass-by-reference function).
For recursive objects, tf is accompanied by the .lapply argument.
By default, .lapply = lapply.
The user may supply a custom lapply()-like function in this argument to use instead.
For example, the perform parallel transformation, the user may supply future.apply::future_lapply.
The supplied function must use the exact same argument convention as lapply, otherwise errors or unexpected behaviour may occur.

Recycling and Coercion

Recycling is not allowed in the modification methods.
So, for example, length(rp) must be equal to the length of the selected subset, or equal to 1.
The user should also take into account the auto-coercion rules of the object's class.
See squarebrackets_coercion for details.