Supported Structures

'squarebrackets' only supports the most common S3 objects, and only those that primarily use square brackets for sub-set operations (hence the name of the package).

One can generally divide the structures supported by 'squarebrackets' along 3 key properties:

atomic vs recursive:
Types logical, integer, double, complex, character, and raw are atomic.
Lists and data.frames are recursive.

dimensionality:
Whether an object is a vector, matrix, array, or data.frame.

mutability:
Base R's S3 classes (except Environments) are generally immutable:
Modifying the object will create a copy (called 'copy-on-modify').
'squabrackets also supports data.tables and mutable_atomic objects, which are mutable:
If desired, one can modify them without copy using pass-by-reference semantics.

Supported Structures

'squarebrackets' supports the following immutable structures:

basic atomic classes
(atomic vectors, matrices, and arrays).

derived atomic vectors
(factors, dates, date-time, etc.).

basic list classes
(recursive vectors, matrices, and arrays).

data.frame
(including the classes tibble, sf-data.frame and sf-tibble).

'squarebrackets' supports the following mutable structures:

mutable_atomic
(mutable_atomic vectors, matrices, and arrays);

data.table
(including the classes tidytable, sf-data.table, and sf-tidytable).

Details

Atomic vs Recursive
The sb_ methods and bind_ implementations provided by 'squarebrackets' work on atomic (see is.atomic) objects.
The sb2_ methods and bind2_ implementations provided by 'squarebrackets' work on recursive (see is.recursive) objects.
The distinction is made between atomic and recursive methods, because R's S3 method dispatch only checks the class of an object.
So the S3 method dispatch makes no distinction, for example, between atomic matrices, and recursive matrices, despite them being completely different kinds of objects (except that they're both 2 dimensional objects).
See squarebrackets_method_dispatch for more details.

Dimensionality
'squarebrackets' supports dimensionless or vector objects (i.e. ndims == 0L).
squarebrackets' supports arrays (see is.array and is.matrix); note that a matrix is simply an array with 2 dimensions.
squarebrackets' also supports data.frame-like objects (see is.data.frame).
Specifically, squarebrackets' supports a wide variety of data.frame classes:
data.frame, data.table, tibble, tidytable;
'squarebrackets' also supports their 'sf'-package compatible counter-parts:
sf-data.frame, sf-data.table, sf-tibble, sf-tidytable.

Dimensionless vectors and dimensional arrays are supported in both their atomic and recursive forms.
Data.frame-like objects, in contrast, only exist in the recursive form (and, as stated, are supported by 'squarebrackets').
Recursive vectors, recursive matrices, and recursive arrays, are collectively referred to as "lists" in the 'squarebrackets' documentation.

Note that the dimensionality of data.frame-like objects is not the same as the dimensionality of (recursive) arrays/matrices.
For any array/matrix x, it holds that length(x) == prod(dim(x)).
But for any data.frame x, it is the case that length(x) == ncol(x).

Mutable vs Immutable
Most S3 objects in base 'R' are immutable:
They have no explicit pass-by-reference semantics.
Environments do have pass-by-reference semantics, but they are not supported by 'squarebrackets'.

'squarebrackets' supports the mutable data.table class (and thus also tidytable, which inherits from data.table).

'squarbrackets' also includes a new class of mutable objects: mutable_atomic objects.
mutable_atomic objects are the same as atomic objects, except they are mutable (hence the name).

The supported immutable structures are:
Atomic and recursive vectors/matrices/arrays, data.frames, and tibbles.

All the functions in the 'squarebrackets' package with the word "set" in their name perform pass-by-reference modification, and thus only work on mutable structures.
All other functions work the same way for both mutable and immutable structures.

Derived Atomic Vector
A special class of objects are the Derived Atomic Vector structures:
structures that are derived from atomic objects, but behave differently.
For example:
Factors, datetime, POSIXct and so on are derived from atomic vectors.
But they have attributes and special methods that make them behave differently.

'squarebrackets' treats derived atomic classes as regular atomic vectors.
There are highly specialized packages to handle objects derived from atomic objects.
For example the 'forcats' package for handling factors, and the 'anytime' package to handle date-time objects.

Not Supported S3 structures
Key-Values storage S3 structures, such as environments, are not supported by 'squarebrackets'.