Regarding Modification By Reference

This help page describes how modification using "pass-by-reference" semantics is handled by the 'squarebrackets' package.
This help page does not explain all the basics of pass-by-reference semantics, as this is treated as prior knowledge.
All functions/methods in the 'squarebrackets' package with the word "set" in the name use pass-by-reference semantics.

Advantages and Disadvantages

The main advantage of pass-by-reference is that much less memory is required to modify objects.
But at least 2 things should be taken into consideration when modifying an object by reference.
First, the coercion rules are slightly different: see squarebrackets_mutable_classes.

Second, if 2 or more variables refer to exactly the same object, changing one variable also changes the other ones.
I.e. the following code,

x <- y <- mutable_atomic(1:16)
sb_set(x, i = 1:6, rp = 8)

modifies not just x, but also y.
This is true even if one of the variables is locked (see bindingIsLocked).
I.e. the following code,

x <- mutable_atomic(1:16)
y <- x
lockBinding("y", environment())
sb_set(x, i = 1:6, rp = 8)

modifies both x and y without error, even though y is a locked constant.

Mutable vs Immutable types

With the exception of environments, most of base R's data types are treated as immutable:
Modifying an object in 'R' will make a copy of the object, something called 'copy-on-modify' semantics.
However, almost any of base R's datatypes can be modified by reference, through R's own 'C' API, or through 'C++' code (like via 'Rcpp'), thus treating these objects as mutable, even though they are not "supposed" to be mutable.
Modifying a base 'R' object by reference can be problematic.
Since 'R', and also most R-packages, treat these objects as immutable, modifying them as-if they are mutable may produce undesired results.

To prevent the issue described above, 'squarebrackets' only supports pass-by-reference semantics on objects that are actually supposed to be mutable.
In relation to this restriction, 'squarebrackets' adds a new class of objects, mutable_atomic, which are simply atomic objects that have the permission to be modified by reference.

Lock Binding

The lockBinding function locks a binding of an object, preventing modification.
'R' also uses locked bindings to prevent modification of objects from package namespaces.
The pass-by-reference semantics of 'squarebrackets' in principle respect this, and disallows modification of objects by reference.

However, lockBinding does not lock the address/pointer of an object, only one particular binding of an object.
This problematic; consider the following example:

x <- mutable_atomic(1:16)
y <- x
lockBinding("y", environment())
sb_set(x, i = 1:6, rp = 8)

In the above code, x and y share the same address, thus pointing to the same memory, yet only y is actually locked.
Since x is not locked, modifying x is allowed.
But since sb_set()/sb2_set() performs modification by reference, y will still be modified, despite being locked.

To remedy the issue as explained above, 'squarebrackets' provides the sb_currentBindings method.
This method can do multiple things. One of which is to look at the address of some object x, find all currently existing bindings in the caller environment sharing the same address as x, and locking all these bindings.

Warning:
The sb_currentBindings method only locks currently existing bindings in the caller environment;
bindings that are created after calling sb_currentBindings will not automatically be locked.
Thus, every time the user creates a new binding of the same object, and the user wishes it to be locked, sb_currentBindings must be called again.

Protected Addresses

To prevent an accidental pass-by-reference modification of objects in the base environment, all addresses of all exported objects in the base environment (baseenv) are stored in the option squarebrackets.protected whenever 'squarebrackets' is loaded, either directly or indirectly.
Needless to say, the user should never touch this option.

Input Variable

Methods/functions that perform in-place modification by reference can be thought of as similar to functions in the style of some_function(x, ...) <- value, in the sense that the variable must actually exist as an actual variable.
Thus things like any of the following,
sb_set(1:10, ...), sb2_set(x$a, ...), or sb_set(base::letters),
will not work.

Views of Lists

Regular lists themselves are not treated as mutable objects by 'squarebrackets'.
However, lists are not actually really objects, merely a (potentially hierarchical) structure of pointers.
Thus, even if a list itself is not treated as mutable, subsets of a list which are themselves mutable classes, are mutable.
For example, if you have a list of data.table objects, the data.tables themselves are mutable.
Therefore, the following will work:

x <- list(
 a = data.table(cola = 1:10, colb = letters[1:10]),
 b = data.table(cola = 11:20, colb = letters[11:20])
)
mypointer <- x$a
sb_set(mypointer, col = "cola", tf = \(x)x^2)

Notice in the above code that mypointer is not a copy of x$a, since they have the same address.
Thus changing mypointer also changes x$a.
In other words: mypointer is what could be called a "view" of x$a.

Protection

Due to the properties given in this help page so far, something like the following will not work:

# letters = base::letters
sb_set(letters, i = 1, rp = "XXX")

The above won't work because:

addresses in baseenv() are protected;

immutable objects are disallowed (you'll have to create a mutable object, which will create a copy of the original, thus keeping the original object safe from modification by reference);

locked bindings are disallowed.

Despite the checks made by this package, the user should never actively try to modify a locked or protected object by reference, as that would defeat the purpose of locking an object.

Examples

# the following code demonstrates how locked bindings,
# such as `base::letters`,
# are being safe-guarded

x <- list(a = base::letters)
mypointer <- x$a # view of a list
address(mypointer) == address(base::letters) # TRUE: point to the same memory
#> [1] TRUE
bindingIsLocked("letters", baseenv()) # base::letters is locked ...
#> [1] TRUE
bindingIsLocked("mypointer", environment()) # ... but this pointer is not!
#> [1] FALSE

if(requireNamespace("tinytest")) {
  tinytest::expect_error(
    sb_set(mypointer, i = 1, rp = "XXX") # this still gives an error though ...
  )
}
#> Loading required namespace: tinytest
#> ----- PASSED      : <-->
#>  call| eval(expr, envir, enclos) 

is.mutable_atomic(mypointer) # ... because it's not of class `mutable_atomic`
#> [1] FALSE

x <- list(
  a = as.mutable_atomic(base::letters) # `as.mutable_atomic()` makes a copy
)
mypointer <- x$a # view of a list
address(mypointer) == address(base::letters) # FALSE: it's a copy
#> [1] FALSE
sb_set(
  mypointer, i = 1, rp = "XXX"  # modifies x, does NOT modify `base::letters`
)
print(x) # x is modified
#> $a
#>  [1] "XXX" "b"   "c"   "d"   "e"   "f"   "g"   "h"   "i"   "j"   "k"   "l"  
#> [13] "m"   "n"   "o"   "p"   "q"   "r"   "s"   "t"   "u"   "v"   "w"   "x"  
#> [25] "y"   "z"  
#> mutable_atomic 
#> typeof:  character 
#> 
base::letters # but this still the same
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m" "n" "o" "p" "q" "r" "s"
#> [20] "t" "u" "v" "w" "x" "y" "z"

# Word of warning:
# the safe-guard in 'squarebrackets' is good, but definitely not perfect.
# Do not actively try to break things; you might actually succeed.