Examples Where the Square Bracket Operators Are Less Convenient

This help page shows some examples where the square bracket operators ( [, [<-) are less than optimally convenient, and how the methods provided by 'squarebrackets' can be helpful in those cases.

Arrays

In order to perform subset operations on some array x with the square brackets operator ([, [<-), one needs to know how many dimensions it has.
I.e. if x has 3 dimensions, one would use:

x[i, j, k, drop = FALSE]

x[i, j, k] <- value

But how would one the use the [ and [<- operators, when number of dimensions of x is not known a-priori?
It’s not impossible, but still rather convoluted.

The methods provided by 'squarebrackets' do not use position-based arguments, and as such work on any arbitrary dimensions without requiring prior knowledge;
see squarebrackets_indx_args for details.

When extracting/removing dimensional sub-set from an array, i.e. x[indx1, indx2], dimnames(x) will be preserved (within the range of the subset of course), but names(x) will be dropped completely.
The methods provided by 'squarebrackets' do not completely remove names(x) when extracting/removing subsets.
After all: why bother giving arrays flat names if they'll be stripped anyway?

Rule-sets for data.frame-like Objects

The data.frame, tibble, data.table, and tidytable classes all inherit from class “data.frame”.
Yet they use different rules regarding the usage of the square bracket operators.
Constantly switching between these rules is annoying, and makes one's code inconsistent.

The methods provided by 'squarebrackets' use the same sub-setting rules for all data.frame inherited classes, thus solving this issue.

The 'squarebrackets' package attempts to keep the data.frame methods as class agnostic as possible, through the class agnostic functionality of the 'collapse' and 'data.table' R-packages.
This attempt to keep data.frame-like classes consistent does, admittedly, result in some oddities in how data.frames are treated by 'squarebrackets', compared to how other classes are treated by 'squarebrackets':

Whole-columns will be auto-coerced when replaced/transformed by sb_mod, but partial columns will not be auto-coerced by default.

The sb_x and sb_rm methods always automatically conserve all attributes (though names and dimensions are adjusted accordingly, of course);
the attributes are not stripped, unlike the other classes.

Giving a data.frame-like object with non-unique column names to the sb_-methods returns an error;
duplicating columns with sb_x will automatically adjust the column names to make them unique.

Annoying Sub-setting By Names

When selecting names for sub-setting, only the first occurrences of the names are selected for the sub-set;
and when un-selecting/removing names for sub-setting, the syntax is very different from selecting names.

The methods provided by 'squarebrackets' uses the same syntax for both selecting and removing sub-sets.
Moreover, selecting/removing sub-sets by names always selects/removes all sub-sets with the given names, not just the first match.

Modification Semantics

‘R’ adheres to copy-on-modify semantics when replacing values using [<-.
But sometimes one would like explicit control when to create a copy, and when to modify using pass-by-reference semantics.

The 'squarebrackets' package provides the sb_mod method to return a copy of an object with modified subsets, and the sb_set method to modify using pass-by-reference semantics.
The idx method can be used in combination with R's own [<- operator for R's default copy-on-modify semantics.

Regarding Other Packages

There are some packages that solve some of these issues.
But using different packages for solving different issues for the same common theme (in this case: solving some inconveniences in the square bracket operators) leads to inconsistent code.
I have not found an R-package that provides a holistic approach to providing alternative methods to the square brackets operators.
Thus, this 'R' package was born.