Method to Coercively Transform (Recursive Subsets of) an Object

This is an S3 Method to completely transform subsets of recursive objects with explicit coercion.

Note that when x is a data.table, one can coercively transform columns by reference (which is more memory efficient), using dt_setcoe.

Usage

sb2_coe(x, ...)

# S3 method for default
sb2_coe(x, i, inv = FALSE, ..., v, .lapply = lapply)

# S3 method for array
sb2_coe(
  x,
  idx = NULL,
  dims = NULL,
  i = NULL,
  inv = FALSE,
  ...,
  v,
  .lapply = lapply
)

# S3 method for data.frame
sb2_coe(x, col = NULL, vars = NULL, inv = FALSE, ..., v)

Arguments

x

a recursive object (list-like or data.frame-like).

...

further arguments passed to or from other methods.

i, col, vars, idx, dims, inv

See squarebrackets_indx_args.
An empty index selection returns the original object unchanged.

v

the coercive transformation function to use.

.lapply

the generic methods use lapply for list- and data.frame-like objects to compute tf() on every list element or dataset column.
The user may supply a custom lapply()-like function in this argument to use instead.
For example, the perform parallel transformation, the user may supply future.apply::future_lapply.
The supplied function must use the exact same argument convention as lapply, otherwise errors or unexpected behaviour may occur.

Value

A copy of the coercively transformed object.

Examples

obj <- data.table::data.table(a = 1:10, b = letters[1:10], c = 11:20, d = factor(letters[1:10]))
str(obj) # notice that columns "a" and "c" are INTEGER (`int`)
#> Classes 'data.table' and 'data.frame':	10 obs. of  4 variables:
#>  $ a: int  1 2 3 4 5 6 7 8 9 10
#>  $ b: chr  "a" "b" "c" "d" ...
#>  $ c: int  11 12 13 14 15 16 17 18 19 20
#>  $ d: Factor w/ 10 levels "a","b","c","d",..: 1 2 3 4 5 6 7 8 9 10
#>  - attr(*, ".internal.selfref")=<externalptr> 
sb2_set(
  obj, filter = ~ (a >= 2) & (c <= 17), vars = is.numeric,
  tf = sqrt # WARNING: sqrt() results in `dbl`, but columns are `int`, so decimals lost
)
#> Warning: 1.414214 (type 'double') at RHS position 1 out-of-range(NA) or truncated (precision lost) when assigning to type 'integer' (column 1 named 'a')
#> Warning: 3.464102 (type 'double') at RHS position 1 out-of-range(NA) or truncated (precision lost) when assigning to type 'integer' (column 3 named 'c')
print(obj)
#>         a      b     c      d
#>     <int> <char> <int> <fctr>
#>  1:     1      a    11      a
#>  2:     1      b     3      b
#>  3:     1      c     3      c
#>  4:     2      d     3      d
#>  5:     2      e     3      e
#>  6:     2      f     4      f
#>  7:     2      g     4      g
#>  8:     8      h    18      h
#>  9:     9      i    19      i
#> 10:    10      j    20      j
obj <- data.table::data.table(a = 1:10, b = letters[1:10], c = 11:20, d = factor(letters[1:10]))
obj <- sb2_coe(obj, vars = is.numeric, v = as.numeric)
str(obj)
#> Classes 'data.table' and 'data.frame':	10 obs. of  4 variables:
#>  $ a: num  1 2 3 4 5 6 7 8 9 10
#>  $ b: chr  "a" "b" "c" "d" ...
#>  $ c: num  11 12 13 14 15 16 17 18 19 20
#>  $ d: Factor w/ 10 levels "a","b","c","d",..: 1 2 3 4 5 6 7 8 9 10
#>  - attr(*, ".internal.selfref")=<externalptr> 
sb2_set(obj,
  filter = ~ (a >= 2) & (c <= 17), vars = is.numeric,
  tf = sqrt # SAFE: coercion performed; so no warnings
) 
print(obj)
#>             a      b         c      d
#>         <num> <char>     <num> <fctr>
#>  1:  1.000000      a 11.000000      a
#>  2:  1.414214      b  3.464102      b
#>  3:  1.732051      c  3.605551      c
#>  4:  2.000000      d  3.741657      d
#>  5:  2.236068      e  3.872983      e
#>  6:  2.449490      f  4.000000      f
#>  7:  2.645751      g  4.123106      g
#>  8:  8.000000      h 18.000000      h
#>  9:  9.000000      i 19.000000      i
#> 10: 10.000000      j 20.000000      j