Method to Coerce and Transform (Recursive Subsets of) an Object

This is an S3 Method to completely transform (a recursive subsets of) an object with explicit coercion.

Given some coercing function v(), the following can be stated about this method.

(1) For atomic objects (vectors, matrices, arrays), this method is equivalent to:

x[] <- v(x)

(2) For factors, this method is equivalent to:

x <- v(x)

(3) For lists, with one or multiple elements specified by argument i, this method is equivalent to:

\(x, i) { x[i] <- lapply(x[i], v); return(x) }

(4) And for data.frame-like objects, with one or multiple columns specified by argument col, this method is equivalent to:

collapse::ftransformv(x, col, v)

Note that when x is a data.table, one can coercively transform columns by reference (which is more memory efficient), using dt_setcoe.

Usage

sb_coe(x, ...)

# S3 method for default
sb_coe(x, v, ...)

# S3 method for factor
sb_coe(x, v, ...)

# S3 method for list
sb_coe(x, i, v, ...)

# S3 method for data.frame
sb_coe(x, col = NULL, vars = NULL, v, ...)

Arguments

x

see squarebrackets_immutable_classes and squarebrackets_mutable_classes.

...

further arguments passed to or from other methods.

v

the coercive transformation function to use.

i, col, vars

See squarebrackets_indx_args.
An empty index selection returns the original object unchanged.

Value

A copy of the coercively transformed object.

Details

When replacing values by reference, the (recursive subset of the) object is never coerced, as that requires making a deep copy; instead, the replacement value is coerced.

For example:
Using sb_set() to replacing/transform one or more values of an integer type (int) mutable_atomic object / data.table column, to become 1.5, will NOT coerce the object or column to a decimal type (dbl); instead, the replacement 1.5 is coerced to the integer 1.

For this reason, the sb_coe() method can be used to coercively transform an object before replacing or transforming values by reference.
See also the Examples section below.

Examples

obj <- data.table::data.table(a = 1:10, b = letters[1:10], c = 11:20, d = factor(letters[1:10]))
str(obj) # notice that columns "a" and "c" are INTEGER (`int`)
#> Classes 'data.table' and 'data.frame':	10 obs. of  4 variables:
#>  $ a: int  1 2 3 4 5 6 7 8 9 10
#>  $ b: chr  "a" "b" "c" "d" ...
#>  $ c: int  11 12 13 14 15 16 17 18 19 20
#>  $ d: Factor w/ 10 levels "a","b","c","d",..: 1 2 3 4 5 6 7 8 9 10
#>  - attr(*, ".internal.selfref")=<externalptr> 
sb_set(
  obj, filter = ~ (a >= 2) & (c <= 17), vars = is.numeric,
  tf = sqrt # WARNING: sqrt() results in `dbl`, but columns are `int`, so decimals lost
)
#> Warning: 1.414214 (type 'double') at RHS position 1 out-of-range(NA) or truncated (precision lost) when assigning to type 'integer' (column 1 named 'a')
#> Warning: 3.464102 (type 'double') at RHS position 1 out-of-range(NA) or truncated (precision lost) when assigning to type 'integer' (column 3 named 'c')
print(obj)
#>         a      b     c      d
#>     <int> <char> <int> <fctr>
#>  1:     1      a    11      a
#>  2:     1      b     3      b
#>  3:     1      c     3      c
#>  4:     2      d     3      d
#>  5:     2      e     3      e
#>  6:     2      f     4      f
#>  7:     2      g     4      g
#>  8:     8      h    18      h
#>  9:     9      i    19      i
#> 10:    10      j    20      j
obj <- data.table::data.table(a = 1:10, b = letters[1:10], c = 11:20, d = factor(letters[1:10]))
obj <- sb_coe(obj, vars = is.numeric, v = as.numeric)
str(obj)
#> Classes 'data.table' and 'data.frame':	10 obs. of  4 variables:
#>  $ a: num  1 2 3 4 5 6 7 8 9 10
#>  $ b: chr  "a" "b" "c" "d" ...
#>  $ c: num  11 12 13 14 15 16 17 18 19 20
#>  $ d: Factor w/ 10 levels "a","b","c","d",..: 1 2 3 4 5 6 7 8 9 10
#>  - attr(*, ".internal.selfref")=<externalptr> 
sb_set(obj,
  filter = ~ (a >= 2) & (c <= 17), vars = is.numeric,
  tf = sqrt # SAFE: coercion performed; so no warnings
) 
print(obj)
#>             a      b         c      d
#>         <num> <char>     <num> <fctr>
#>  1:  1.000000      a 11.000000      a
#>  2:  1.414214      b  3.464102      b
#>  3:  1.732051      c  3.605551      c
#>  4:  2.000000      d  3.741657      d
#>  5:  2.236068      e  3.872983      e
#>  6:  2.449490      f  4.000000      f
#>  7:  2.645751      g  4.123106      g
#>  8:  8.000000      h 18.000000      h
#>  9:  9.000000      i 19.000000      i
#> 10: 10.000000      j 20.000000      j