Methods to Insert New Values Before or After an Index Along a Dimension

The sb_before()/sb2_before() method inserts new values before some position along a dimension.
The sb_after()/sb2_after() method inserts new values after some position along a dimension.

Use sb_before(x, ...)/sb_after(x, ...) if x is a non-recursive object (i.e. atomic or factor).
Use sb2_before(x, ...)/sb2_after(x, ...) if x is a recursive object (i.e. list or data.frame-like).

sb2_before.array() and sb2_after.array() use a altered version of abind::abind (see references below), which has been modified to work on recursive arrays (i.e. dimensional lists).

Usage

sb_before(x, ...)

sb_after(x, ...)

# S3 method for default
sb_before(x, new, pos = 1, .attr = NULL, ...)

# S3 method for default
sb_after(x, new, pos = length(x), .attr = NULL, ...)

# S3 method for array
sb_before(x, new, margin, pos = 1, .attr = NULL, ...)

# S3 method for array
sb_after(x, new, margin, pos = dim(x)[margin], .attr = NULL, ...)

# S3 method for factor
sb_before(x, new, pos = 1, .attr = NULL, ...)

# S3 method for factor
sb_after(x, new, pos = length(x), .attr = NULL, ...)

sb2_before(x, ...)

sb2_after(x, ...)

# S3 method for default
sb2_before(x, new, pos = 1, .attr = NULL, ...)

# S3 method for default
sb2_after(x, new, pos = length(x), .attr = NULL, ...)

# S3 method for array
sb2_before(x, new, margin, pos = 1, .attr = NULL, ...)

# S3 method for array
sb2_after(x, new, margin, pos = dim(x)[margin], .attr = NULL, ...)

# S3 method for data.frame
sb2_before(x, new, margin, pos = 1, .attr = NULL, ...)

# S3 method for data.frame
sb2_after(x, new, margin, pos = collapse::fdim(x)[margin], .attr = NULL, ...)

Arguments

x

see squarebrackets_immutable_classes and squarebrackets_mutable_classes.

...

further arguments passed to or from other methods.

new

the new value(s). The type of object depends on x:

For atomic objects, new can be any atomic object. However, if one wished the added values in new to be named, ensure new is the same type of object as x. For example: use matrix with column names for new when appending/inserting columns to matrix x.

For factors, new must be a factor.

For lists, new must be a (possible named) list.

For data.frame-like objects, new must be a data.frame.

pos

a strictly positive single integer scalar (so no duplicates), giving the position along the dimension (specified in margin), before or after which the new values are added.

.attr

a list, giving additional potentially missing attributes to be added to the returned object.
By default, concatenation strips attributes, since the attributes of x and new may not be equal or even compatible.
In the attr argument, the attributes of the merged object can be specified.
Only attributes that are actually missing AFTER insertion will be added, thus preventing overwriting existing attributes like names.
One may, for example, specify .attr = sb_a(x) or .attr = sb_a(new).
If NULL (the default), no attributes will be added post-insert.
If speed is important, NULL is the best option (but then attributes won't be preserved).

margin

a single scalar, giving the dimension along which to add new values.

Value

Returns a copy of the appended object.

References

Plate T, Heiberger R (2016). abind: Combine Multidimensional Arrays. R package version 1.4-5, https://CRAN.R-project.org/package=abind.

Examples

# atomic objects ====

x <- matrix(1:20 , ncol = 4)
print(x)
#>      [,1] [,2] [,3] [,4]
#> [1,]    1    6   11   16
#> [2,]    2    7   12   17
#> [3,]    3    8   13   18
#> [4,]    4    9   14   19
#> [5,]    5   10   15   20
new <- -1 * x
sb_before(x, new, 1)
#>       [,1] [,2] [,3] [,4]
#>  [1,]   -1   -6  -11  -16
#>  [2,]   -2   -7  -12  -17
#>  [3,]   -3   -8  -13  -18
#>  [4,]   -4   -9  -14  -19
#>  [5,]   -5  -10  -15  -20
#>  [6,]    1    6   11   16
#>  [7,]    2    7   12   17
#>  [8,]    3    8   13   18
#>  [9,]    4    9   14   19
#> [10,]    5   10   15   20
sb_after(x, new, 1)
#>       [,1] [,2] [,3] [,4]
#>  [1,]    1    6   11   16
#>  [2,]    2    7   12   17
#>  [3,]    3    8   13   18
#>  [4,]    4    9   14   19
#>  [5,]    5   10   15   20
#>  [6,]   -1   -6  -11  -16
#>  [7,]   -2   -7  -12  -17
#>  [8,]   -3   -8  -13  -18
#>  [9,]   -4   -9  -14  -19
#> [10,]   -5  -10  -15  -20
sb_before(x, new, 2)
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
#> [1,]   -1   -6  -11  -16    1    6   11   16
#> [2,]   -2   -7  -12  -17    2    7   12   17
#> [3,]   -3   -8  -13  -18    3    8   13   18
#> [4,]   -4   -9  -14  -19    4    9   14   19
#> [5,]   -5  -10  -15  -20    5   10   15   20
sb_after(x, new, 2)
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
#> [1,]    1    6   11   16   -1   -6  -11  -16
#> [2,]    2    7   12   17   -2   -7  -12  -17
#> [3,]    3    8   13   18   -3   -8  -13  -18
#> [4,]    4    9   14   19   -4   -9  -14  -19
#> [5,]    5   10   15   20   -5  -10  -15  -20

#############################################################################

# factors ====

x <- factor(letters)
new <- factor("foo")
sb_before(x, new)
#>  [1] foo a   b   c   d   e   f   g   h   i   j   k   l   m   n   o   p   q   r  
#> [20] s   t   u   v   w   x   y   z  
#> Levels: foo a b c d e f g h i j k l m n o p q r s t u v w x y z
sb_after(x, new)
#>  [1] a   b   c   d   e   f   g   h   i   j   k   l   m   n   o   p   q   r   s  
#> [20] t   u   v   w   x   y   z   foo
#> Levels: a b c d e f g h i j k l m n o p q r s t u v w x y z foo

#############################################################################

# lists ====

x <- as.list(1:5)
new <- lapply(x, \(x)x*-1)
print(x)
#> [[1]]
#> [1] 1
#> 
#> [[2]]
#> [1] 2
#> 
#> [[3]]
#> [1] 3
#> 
#> [[4]]
#> [1] 4
#> 
#> [[5]]
#> [1] 5
#> 
sb2_before(x, new)
#> [[1]]
#> [1] -1
#> 
#> [[2]]
#> [1] -2
#> 
#> [[3]]
#> [1] -3
#> 
#> [[4]]
#> [1] -4
#> 
#> [[5]]
#> [1] -5
#> 
#> [[6]]
#> [1] 1
#> 
#> [[7]]
#> [1] 2
#> 
#> [[8]]
#> [1] 3
#> 
#> [[9]]
#> [1] 4
#> 
#> [[10]]
#> [1] 5
#> 
sb2_after(x, new)
#> [[1]]
#> [1] 1
#> 
#> [[2]]
#> [1] 2
#> 
#> [[3]]
#> [1] 3
#> 
#> [[4]]
#> [1] 4
#> 
#> [[5]]
#> [1] 5
#> 
#> [[6]]
#> [1] -1
#> 
#> [[7]]
#> [1] -2
#> 
#> [[8]]
#> [1] -3
#> 
#> [[9]]
#> [1] -4
#> 
#> [[10]]
#> [1] -5
#> 

#############################################################################

# recursive arrays / dimensional lists ====

x <- matrix(c(as.list(1:20), as.list(letters[1:20])) , ncol = 8) |> t()
print(x)
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,] 1    2    3    4    5   
#> [2,] 6    7    8    9    10  
#> [3,] 11   12   13   14   15  
#> [4,] 16   17   18   19   20  
#> [5,] "a"  "b"  "c"  "d"  "e" 
#> [6,] "f"  "g"  "h"  "i"  "j" 
#> [7,] "k"  "l"  "m"  "n"  "o" 
#> [8,] "p"  "q"  "r"  "s"  "t" 
new <- matrix(c(as.list(-1:-20), as.list(letters[26:7])) , ncol = 8) |> t()
sb2_before(x, new, 1)
#>       [,1] [,2] [,3] [,4] [,5]
#>  [1,] -1   -2   -3   -4   -5  
#>  [2,] -6   -7   -8   -9   -10 
#>  [3,] -11  -12  -13  -14  -15 
#>  [4,] -16  -17  -18  -19  -20 
#>  [5,] "z"  "y"  "x"  "w"  "v" 
#>  [6,] "u"  "t"  "s"  "r"  "q" 
#>  [7,] "p"  "o"  "n"  "m"  "l" 
#>  [8,] "k"  "j"  "i"  "h"  "g" 
#>  [9,] 1    2    3    4    5   
#> [10,] 6    7    8    9    10  
#> [11,] 11   12   13   14   15  
#> [12,] 16   17   18   19   20  
#> [13,] "a"  "b"  "c"  "d"  "e" 
#> [14,] "f"  "g"  "h"  "i"  "j" 
#> [15,] "k"  "l"  "m"  "n"  "o" 
#> [16,] "p"  "q"  "r"  "s"  "t" 
sb2_after(x, new, 1)
#>       [,1] [,2] [,3] [,4] [,5]
#>  [1,] 1    2    3    4    5   
#>  [2,] 6    7    8    9    10  
#>  [3,] 11   12   13   14   15  
#>  [4,] 16   17   18   19   20  
#>  [5,] "a"  "b"  "c"  "d"  "e" 
#>  [6,] "f"  "g"  "h"  "i"  "j" 
#>  [7,] "k"  "l"  "m"  "n"  "o" 
#>  [8,] "p"  "q"  "r"  "s"  "t" 
#>  [9,] -1   -2   -3   -4   -5  
#> [10,] -6   -7   -8   -9   -10 
#> [11,] -11  -12  -13  -14  -15 
#> [12,] -16  -17  -18  -19  -20 
#> [13,] "z"  "y"  "x"  "w"  "v" 
#> [14,] "u"  "t"  "s"  "r"  "q" 
#> [15,] "p"  "o"  "n"  "m"  "l" 
#> [16,] "k"  "j"  "i"  "h"  "g" 
sb2_before(x, new, 2)
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,] -1   -2   -3   -4   -5   1    2    3    4    5    
#> [2,] -6   -7   -8   -9   -10  6    7    8    9    10   
#> [3,] -11  -12  -13  -14  -15  11   12   13   14   15   
#> [4,] -16  -17  -18  -19  -20  16   17   18   19   20   
#> [5,] "z"  "y"  "x"  "w"  "v"  "a"  "b"  "c"  "d"  "e"  
#> [6,] "u"  "t"  "s"  "r"  "q"  "f"  "g"  "h"  "i"  "j"  
#> [7,] "p"  "o"  "n"  "m"  "l"  "k"  "l"  "m"  "n"  "o"  
#> [8,] "k"  "j"  "i"  "h"  "g"  "p"  "q"  "r"  "s"  "t"  
sb2_after(x, new, 2)
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,] 1    2    3    4    5    -1   -2   -3   -4   -5   
#> [2,] 6    7    8    9    10   -6   -7   -8   -9   -10  
#> [3,] 11   12   13   14   15   -11  -12  -13  -14  -15  
#> [4,] 16   17   18   19   20   -16  -17  -18  -19  -20  
#> [5,] "a"  "b"  "c"  "d"  "e"  "z"  "y"  "x"  "w"  "v"  
#> [6,] "f"  "g"  "h"  "i"  "j"  "u"  "t"  "s"  "r"  "q"  
#> [7,] "k"  "l"  "m"  "n"  "o"  "p"  "o"  "n"  "m"  "l"  
#> [8,] "p"  "q"  "r"  "s"  "t"  "k"  "j"  "i"  "h"  "g"  

#############################################################################

# data.frame-like objects ====

x <- data.frame(a = 1:10, b = letters[1:10], c = 11:20, d = factor(letters[1:10]))
new <- data.frame(e = 101:110)
sb2_before(x, new, 2)
#>      e  a b  c d
#> 1  101  1 a 11 a
#> 2  102  2 b 12 b
#> 3  103  3 c 13 c
#> 4  104  4 d 14 d
#> 5  105  5 e 15 e
#> 6  106  6 f 16 f
#> 7  107  7 g 17 g
#> 8  108  8 h 18 h
#> 9  109  9 i 19 i
#> 10 110 10 j 20 j
sb2_after(x, new, 2)
#>     a b  c d   e
#> 1   1 a 11 a 101
#> 2   2 b 12 b 102
#> 3   3 c 13 c 103
#> 4   4 d 14 d 104
#> 5   5 e 15 e 105
#> 6   6 f 16 f 106
#> 7   7 g 17 g 107
#> 8   8 h 18 h 108
#> 9   9 i 19 i 109
#> 10 10 j 20 j 110
new <- x[1,]
sb2_before(x, new, 1)
#>     a b  c d
#> 1   1 a 11 a
#> 2   1 a 11 a
#> 3   2 b 12 b
#> 4   3 c 13 c
#> 5   4 d 14 d
#> 6   5 e 15 e
#> 7   6 f 16 f
#> 8   7 g 17 g
#> 9   8 h 18 h
#> 10  9 i 19 i
#> 11 10 j 20 j
sb2_after(x, new, 1)
#>     a b  c d
#> 1   1 a 11 a
#> 2   2 b 12 b
#> 3   3 c 13 c
#> 4   4 d 14 d
#> 5   5 e 15 e
#> 6   6 f 16 f
#> 7   7 g 17 g
#> 8   8 h 18 h
#> 9   9 i 19 i
#> 10 10 j 20 j
#> 11  1 a 11 a