Access Recursive Subsets

The sb_rec() method allows the user to access recursive subsets of lists.

The sb_rec() method also allows replacing or transforming a recursive subset of a list, using R's default in-place semantics, by specifying the rp argument.

Usage

sb_rec(lst, rec, rp)

Arguments

lst

a list, or list-like object.

rec

a vector of length p, such that lst[[rec]] is equivalent to lst[[ rec[1] ]]...[[ rec[p] ]], providing all but the final indexing results in a list.
When on a certain subset level of a nested list, multiple subsets with the same name exist, only the first one will be selected when performing recursive indexing by name, due to the recursive nature of this type of subsetting.

rp

optional. If specified, performs lst[[rec]] <- rp, using R's default in-place semantics.

Value

If rp is not specified: Returns the recursive subset.

If rp is specified: Returns nothing, but replaces the recursive subset with rp, using R's default in-place semantics.

Examples

lst <- list(
  A = list(
    A = list(A = "AAA", B = "AAB"),
    A = list(A  = "AA2A", B = "AA2B"),
    B = list(A = "ABA", B = "ABB")
  ),
  B = list(
    A = list(A = "BAA", B = "BAB"),
    B = list(A = "BBA", B = "BBB")
  )
)

#############################################################################

# access recursive subsets ====

sb_rec(lst, c(1,2,2)) # this gives "AA2B"
#> [1] "AA2B"
sb_rec(lst, c("A", "B", "B")) # this gives "ABB"
#> [1] "ABB"
sb_rec(lst, c(2,2,1)) # this gives "BBA"
#> [1] "BBA"
sb_rec(lst, c("B", "B", "A")) # this gives "BBA"
#> [1] "BBA"

#############################################################################

# replace with R's default in-place semantics ====

# replace "AAB" using R's default in-place semantics:
sb_rec(
  lst, c("A", "A", "B"),
  rp = "THIS IS REPLACED WITH IN-PLACE SEMANTICS"
)
print(lst)
#> $A
#> $A$A
#> $A$A$A
#> [1] "AAA"
#> 
#> $A$A$B
#> [1] "THIS IS REPLACED WITH IN-PLACE SEMANTICS"
#> 
#> 
#> $A$A
#> $A$A$A
#> [1] "AA2A"
#> 
#> $A$A$B
#> [1] "AA2B"
#> 
#> 
#> $A$B
#> $A$B$A
#> [1] "ABA"
#> 
#> $A$B$B
#> [1] "ABB"
#> 
#> 
#> 
#> $B
#> $B$A
#> $B$A$A
#> [1] "BAA"
#> 
#> $B$A$B
#> [1] "BAB"
#> 
#> 
#> $B$B
#> $B$B$A
#> [1] "BBA"
#> 
#> $B$B$B
#> [1] "BBB"
#> 
#> 
#> 

#############################################################################

# Modify View of List By Reference ====

x <- list(
 a = data.table::data.table(cola = 1:10, colb = letters[1:10]),
 b = data.table::data.table(cola = 11:20, colb = letters[11:20])
)
print(x)
#> $a
#>      cola   colb
#>     <int> <char>
#>  1:     1      a
#>  2:     2      b
#>  3:     3      c
#>  4:     4      d
#>  5:     5      e
#>  6:     6      f
#>  7:     7      g
#>  8:     8      h
#>  9:     9      i
#> 10:    10      j
#> 
#> $b
#>      cola   colb
#>     <int> <char>
#>  1:    11      k
#>  2:    12      l
#>  3:    13      m
#>  4:    14      n
#>  5:    15      o
#>  6:    16      p
#>  7:    17      q
#>  8:    18      r
#>  9:    19      s
#> 10:    20      t
#> 
mypointer <- sb_rec(x, "a")
address(mypointer) == address(x$a) # they are the same
#> [1] TRUE
sb_set(mypointer, col = "cola", tf = \(x)x^2)
print(x) # notice x has been changed
#> $a
#>      cola   colb
#>     <num> <char>
#>  1:     1      a
#>  2:     4      b
#>  3:     9      c
#>  4:    16      d
#>  5:    25      e
#>  6:    36      f
#>  7:    49      g
#>  8:    64      h
#>  9:    81      i
#> 10:   100      j
#> 
#> $b
#>      cola   colb
#>     <int> <char>
#>  1:    11      k
#>  2:    12      l
#>  3:    13      m
#>  4:    14      n
#>  5:    15      o
#>  6:    16      p
#>  7:    17      q
#>  8:    18      r
#>  9:    19      s
#> 10:    20      t
#>