Method to Un-Select/Remove Subsets of an Object

This is an S3 Method to un-select/remove subsets from an object.

Usage

sb_rm(x, ...)

# S3 method for default
sb_rm(x, i, ..., rat = FALSE, chkdup = TRUE)

# S3 method for matrix
sb_rm(x, row = NULL, col = NULL, i = NULL, ..., rat = FALSE, chkdup = TRUE)

# S3 method for array
sb_rm(
  x,
  idx = NULL,
  dims = NULL,
  rcl = NULL,
  i = NULL,
  ...,
  rat = FALSE,
  chkdup = TRUE
)

# S3 method for factor
sb_rm(x, i = NULL, lvl = NULL, drop = FALSE, ..., rat = FALSE, chkdup = TRUE)

# S3 method for list
sb_rm(x, i, drop = FALSE, ..., rat = FALSE, chkdup = TRUE)

# S3 method for data.frame
sb_rm(
  x,
  row = NULL,
  col = NULL,
  filter = NULL,
  vars = NULL,
  ...,
  chkdup = TRUE
)

Arguments

x

see squarebrackets_immutable_classes and squarebrackets_mutable_classes.

...

further arguments passed to or from other methods.

i, lvl, row, col, idx, dims, rcl, filter, vars

See squarebrackets_indx_args.
An empty index selection results in nothing being removed, and the entire object is returned.

rat

Boolean, indicating if attributes should be returned with the sub-setted object.
See Details section for more info.

chkdup

see squarebrackets_duplicates.

drop

Boolean.

For factors: If drop = TRUE, unused levels are dropped, if drop = FALSE they are not dropped.

For lists: if drop = TRUE, selecting a single element will give the simplified result, like using [[]]. If drop = FALSE, a list is always returned regardless of the number of elements.

Value

A copy of the sub-setted object.

Details

One the rat argument
Most [ - methods strip most (but not all) attributes.
If rat = FALSE, chkdup = TRUE, this default behaviour is preserved, for compatibility with special classes. This is the fastest option.
If rat = TRUE, attributes from x missing after sub-setting are re-assigned to x. Already existing attributes after sub-setting will not be overwritten.
There is no rat argument for data.frame-like object: their attributes will always be preserved.
NOTE: In the following situations, the rat argument will be ignored, as the attributes necessarily have to be dropped:

when x is a list, AND drop = TRUE, AND a single element is selected.

when x is a matrix or array, and sub-setting is done through the i argument.

Examples

# atomic objects ====

obj <- matrix(1:16, ncol = 4)
colnames(obj) <- c("a", "b", "c", "a")
print(obj)
#>      a b  c  a
#> [1,] 1 5  9 13
#> [2,] 2 6 10 14
#> [3,] 3 7 11 15
#> [4,] 4 8 12 16
sb_rm(obj, 1:3, 1:3)
#>       a
#> [1,] 16
# above is equivalent to  obj[-1:-3, -1:-3, drop = FALSE]
sb_rm(obj, i = \(x)x>5)
#> [1] 1 2 3 4 5
# above is equivalent to  obj[!obj > 5]
sb_rm(obj, col = "a")
#>      b  c
#> [1,] 5  9
#> [2,] 6 10
#> [3,] 7 11
#> [4,] 8 12
# above is equivalent to  obj[, which(!colnames(obj) %in% "a")]

obj <- array(1:64, c(4,4,3))
print(obj)
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    1    5    9   13
#> [2,]    2    6   10   14
#> [3,]    3    7   11   15
#> [4,]    4    8   12   16
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   17   21   25   29
#> [2,]   18   22   26   30
#> [3,]   19   23   27   31
#> [4,]   20   24   28   32
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   33   37   41   45
#> [2,]   34   38   42   46
#> [3,]   35   39   43   47
#> [4,]   36   40   44   48
#> 
sb_rm(obj, n(1, c(1, 3)), c(1, 3))
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   18   22   26   30
#> [2,]   19   23   27   31
#> [3,]   20   24   28   32
#> 
sb_rm(obj, rcl = n(1, NULL, c(1, 3)))
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   18   22   26   30
#> [2,]   19   23   27   31
#> [3,]   20   24   28   32
#> 
# above 2 lines are equivalent to obj[-1, c(-1, -3), drop = FALSE]
sb_rm(obj, i = \(x)x>5)
#> [1] 1 2 3 4 5
# above is equivalent to obj[!obj > 5]

#############################################################################

# lists ====

obj <- list(a = 1:10, b = letters[1:11], c = 11:20)
print(obj)
#> $a
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> $b
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k"
#> 
#> $c
#>  [1] 11 12 13 14 15 16 17 18 19 20
#> 
sb_rm(obj, "a")
#> $b
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k"
#> 
#> $c
#>  [1] 11 12 13 14 15 16 17 18 19 20
#> 
# above is equivalent to obj[which(!names(obj) %in% "a")]
sb_rm(obj, 1) # obj[-1]
#> $b
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k"
#> 
#> $c
#>  [1] 11 12 13 14 15 16 17 18 19 20
#> 
sb_rm(obj, 1:2)
#> $c
#>  [1] 11 12 13 14 15 16 17 18 19 20
#> 
# above is equivalent to obj[seq_len(length(obj))[-1:-2]]
sb_rm(obj, is.numeric, drop = TRUE)
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k"
# above is equivalent to obj[[!sapply(obj, is.numeric)]] IF this returns a single element
obj <- list(a = 1:10, b = letters[1:11], c = letters)
sb_rm(obj, is.numeric)
#> $b
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k"
#> 
#> $c
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m" "n" "o" "p" "q" "r" "s"
#> [20] "t" "u" "v" "w" "x" "y" "z"
#> 
# above is equivalent to obj[!sapply(obj, is.numeric)] # this time singular brackets?
# for recusive indexing, see sb_rec()

#############################################################################

# factors ====

obj <- factor(rep(letters[1:5], 2))
sb_rm(obj, lvl = "a")
#> [1] b c d e b c d e
#> Levels: a b c d e
# above is equivalent to obj[which(!obj %in% "a")]

#############################################################################

# data.frame-like objects ====

obj <- data.frame(a = 1:10, b = letters[1:10], c = 11:20, d = factor(letters[1:10]))
print(obj)
#>     a b  c d
#> 1   1 a 11 a
#> 2   2 b 12 b
#> 3   3 c 13 c
#> 4   4 d 14 d
#> 5   5 e 15 e
#> 6   6 f 16 f
#> 7   7 g 17 g
#> 8   8 h 18 h
#> 9   9 i 19 i
#> 10 10 j 20 j
sb_rm(obj, 1:3, 1:3)
#>   d
#> 1 d
#> 2 e
#> 3 f
#> 4 g
#> 5 h
#> 6 i
#> 7 j
# above is equivalent to obj[-1:-3, -1:-3, drop = FALSE]
sb_rm(obj, filter = ~ (a > 5) & (c < 19), vars = is.numeric)
#>   b d
#> 1 a a
#> 2 b b
#> 3 c c
#> 4 d d
#> 5 e e
#> 6 i i
#> 7 j j