Method to Change the Names of a Mutable Object By Reference

This is an S3 Method to rename a supported mutable object using pass-by-reference semantics.

This method takes extra care not to modify any objects that happen to share the same address as the (dim)names of x.
I.e. the following code:

x <- mutable_atomic(1:26)
names(x) <- base::letters
y <- x
sb_setRename(x, newnames = rev(names(x)))

will not modify base::letters, even though names(x) shared the same address.
Thus, sb_setRename() can be used safely without fearing such accidents.

Use sb_setRename(x, ...) if x is a non-recursive object (i.e. mutable_atomic).
Use sb2_setRename(x, ...) if x is a recursive object (i.e. data.table).

Usage

sb_setRename(x, ...)

# S3 method for default
sb_setRename(x, newnames, ...)

# S3 method for array
sb_setRename(x, newnames, newdimnames, ...)

sb2_setRename(x, ...)

# S3 method for data.table
sb2_setRename(x, old, new, skip_absent = FALSE, ...)

Arguments

x

a variable belonging to one of the supported mutable classes.

...

further arguments passed to or from other methods.

newnames

atomic character vector giving the new names.
Specifying NULL will remove the names.

newdimnames

a list of the same length as dim(x).
The first element of the list corresponds to the first dimension, the second element to the second dimension, and so on.
The components of the list can be either NULL, or a character vector with the same length as the corresponding dimension.
Instead of a list, simply NULL can be specified, which will remove the dimnames completely.

old

the old column names

new

the new column names, in the same order as old

skip_absent

Skip items in old that are missing (i.e. absent) in names(x).
Default FALSE halts with error if any are missing.

Value

Returns: VOID. This method modifies the object by reference.

Do not use assignment like names(x) <- sb_setRename(x, ...).

Since this function returns void, you'll just get NULL.

Examples

 
# mutable atomic vector ====
x <- y <- mutable_atomic(1:10, names = letters[1:10])
sb_setRename(x, rev(letters[1:10]))
print(x)
#>  j  i  h  g  f  e  d  c  b  a 
#>  1  2  3  4  5  6  7  8  9 10 
#> mutable_atomic 
#> typeof:  integer 

################################################################################

# mutable atomic matrix ====
x <- mutable_atomic(
  1:20, dim = c(5, 4), dimnames = n(letters[1:5], letters[1:4])
)
print(x)
#>   a  b  c  d
#> a 1  6 11 16
#> b 2  7 12 17
#> c 3  8 13 18
#> d 4  9 14 19
#> e 5 10 15 20
#> mutable_atomic 
#> typeof:  integer 
sb_setRename(
  x,
  newdimnames = lapply(dimnames(x), rev)
)
print(x)
#>   d  c  b  a
#> e 1  6 11 16
#> d 2  7 12 17
#> c 3  8 13 18
#> b 4  9 14 19
#> a 5 10 15 20
#> mutable_atomic 
#> typeof:  integer 

x <- mutable_atomic(
   1:20, letters[1:20], dim = c(5, 4), dimnames = n(letters[1:5], letters[1:4])
)
print(x)
#>   a  b  c  d
#> a 1  6 11 16
#> b 2  7 12 17
#> c 3  8 13 18
#> d 4  9 14 19
#> e 5 10 15 20
#> attr(,"names")
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m" "n" "o" "p" "q" "r" "s"
#> [20] "t"
#> mutable_atomic 
#> typeof:  integer 
sb_setRename(
  x, newnames = rev(names(x)),
  newdimnames = lapply(dimnames(x), rev)
)
print(x)
#>   d  c  b  a
#> e 1  6 11 16
#> d 2  7 12 17
#> c 3  8 13 18
#> b 4  9 14 19
#> a 5 10 15 20
#> attr(,"names")
#>  [1] "t" "s" "r" "q" "p" "o" "n" "m" "l" "k" "j" "i" "h" "g" "f" "e" "d" "c" "b"
#> [20] "a"
#> mutable_atomic 
#> typeof:  integer 

################################################################################

# data.table ====

x <- data.table::data.table(
  a = 1:20,
  b = letters[1:20]
)
print(x)
#>         a      b
#>     <int> <char>
#>  1:     1      a
#>  2:     2      b
#>  3:     3      c
#>  4:     4      d
#>  5:     5      e
#>  6:     6      f
#>  7:     7      g
#>  8:     8      h
#>  9:     9      i
#> 10:    10      j
#> 11:    11      k
#> 12:    12      l
#> 13:    13      m
#> 14:    14      n
#> 15:    15      o
#> 16:    16      p
#> 17:    17      q
#> 18:    18      r
#> 19:    19      s
#> 20:    20      t
#>         a      b
sb2_setRename(x, old = names(x), new = rev(names(x)))
print(x)
#>         b      a
#>     <int> <char>
#>  1:     1      a
#>  2:     2      b
#>  3:     3      c
#>  4:     4      d
#>  5:     5      e
#>  6:     6      f
#>  7:     7      g
#>  8:     8      h
#>  9:     9      i
#> 10:    10      j
#> 11:    11      k
#> 12:    12      l
#> 13:    13      m
#> 14:    14      n
#> 15:    15      o
#> 16:    16      p
#> 17:    17      q
#> 18:    18      r
#> 19:    19      s
#> 20:    20      t
#>         b      a