Specialized Sub-setting Functions

The sb_a() function subsets extracts one or more attributes from an object.

The sb_str() function subsets characters of single string, or replace a subset of the characters of a single string with the subsets of the characters of another string. In both cases, a single string is treated as a iterable vector, where each single character in a string is a single element. The sb_str() function is considerably faster than doing the equivalent operation in base 'R' or even 'stringi'.

Usage

sb_str(str, ind, rp.str, rp.ind)

sb_a(x, a = NULL)

Arguments

str

a single string.

ind

an integer vector, giving the positions of the string to subset.

rp.str, rp.ind

similar to str and ind, respectively.
If not specified, sb_str() will perform something like
str[ind]
treating str as an iterable vector.
If these ARE specified, sb_str() will perform something like
str[ind] <- rp.str[rp.ind]
treating str and rp.str as iterable vectors.

x

an object

a

a character vector of attribute names. If NULL (default), ALL attributes are returned.

Value

The sub-setted object.

Examples

x <- matrix(1:10, ncol = 2)
colnames(x) <- c("a", "b")
attr(x, "test") <- "test"
sb_a(x, "test")
#> $test
#> [1] "test"
#> 
sb_a(x)
#> $dim
#> [1] 5 2
#> 
#> $dimnames
#> $dimnames[[1]]
#> NULL
#> 
#> $dimnames[[2]]
#> [1] "a" "b"
#> 
#> 
#> $test
#> [1] "test"
#> 

x <- "hello"
sb_str(x, 5:1) # this gives "olleh"
#> [1] "olleh"
sb_str(x, c(1:5, 5)) # this gives "helloo"
#> [1] "helloo"
sb_str(x, c(2:5)) # this gives "ello"
#> [1] "ello"
sb_str(x, seq(1, 5, by = 2)) # this gives "hlo"
#> [1] "hlo"
sb_str(x, 1:4, "world", 1:4) # this gives "worlo"
#> [1] "worlo"