Recursive Sequence Generator

This is a recursive sequence generator. The function is essentially a highly generalized version of a Fibonacci sequence generator. One can change the initial values, the window size, and even the window function used.
This function assumes only the following about the sequence being generated:

The sequence consists of real numbers (i.e. class integer or class double).

The window size is the same for all iterations.

The window function is the same for all iterations.

The sequence grows until a vector of length n is achieved.

Usage

seq_rec(inits = c(0, 1), n = 10L, f = sum)

Arguments

inits

a numeric (double or integer) vector giving the initial values.
Any numbers are allowed, even negative and/or fractional numbers.
Note that numbers given must give valid results when passed to function f().
IMPORTANT: The length of inits determines the window size w.
For a regular Fibonacci, inits = 0:1, which of course means a window size of w = 2.

n

a single integer, giving the size of the numeric vector to generate.
NOTE: it must hold that n is larger than or equal to the window size w.
The window size is equal to w = length(inits).

f

the function to be used on the last w numbers to generate the next number of the sequence at each iteration.
This must be a function that takes as input a numeric vector, and returns a single numeric value.
For a regular Fibonacci sequence, this would be either:
f = sum,
or (since window size is 2) f = \(x) x[2] + x[1]

Value

A sequence of numbers.

Details

The default values of the arguments give the first 10 numbers of a regular Fibonacci sequence.
See examples for several number series created with this function.
This function is written in C++ using Rcpp for better performance.

Examples

seq_rec() # by default gives Fibonacci numbers
#>  [1]  0  1  1  2  3  5  8 13 21 34
seq_rec(0:3, 10L, sum) # a weird shifted version of Fibonacci
#>  [1]   0   1   2   3   6  12  23  44  85 164
seq_rec(inits = 2:1) # Lucas numbers
#>  [1]  2  1  3  4  7 11 18 29 47 76
c(1, seq_rec(c(1, 2), f = prod)) # Multiplicative Fibonacci
#>  [1]           1           1           2           2           4           8
#>  [7]          32         256        8192     2097152 17179869184
seq_rec(f = \(x) 2 * x[2] + x[1]) # Pell numbers
#>  [1]   0   1   2   5  12  29  70 169 408 985
seq_rec(inits = c(1, 0), f = \(x) 2 * x[1]) # see https://oeis.org/A077957
#>  [1]  1  0  2  0  4  0  8  0 16  0
seq_rec(f = \(x) x[2] + 2 * x[1]) # Jacobsthal numbers
#>  [1]   0   1   1   3   5  11  21  43  85 171
seq_rec(c(1, 1, 1), f = \(x) x[1] + x[2]) # Padovan sequence
#>  [1] 1 1 1 2 2 3 4 5 7 9
seq_rec(c(3, 0, 2), f = \(x) x[1] + x[2]) # Perrin numbers
#>  [1]  3  0  2  3  2  5  5  7 10 12
seq_rec(c(0, 1, 3), f = \(x) 3 * x[3] - 3 * x[2] + x[1]) # Triangular numbers
#>  [1]  0  1  3  6 10 15 21 28 36 45