Generate Recursive Sequence Through Repeated Arithmetic Infix Operations

This is a recursive sequence generator.
The function is essentially a highly generalized version of a Fibonacci sequence generator.

Starting with 2 initial values, each next value i is generated by either one of 2 formulas:

x[i] = (s[1] + m[1] * x[i-1]) %inop% (s[2] + m[2] * x[i-2])

x[i] = (m[1] * (x[i-1] + s[1])) %inop% (m[2] * (x[i-2] + s[2]))

where %inop% is the arithmetic infix operator chosen,
and m and s are each a numeric vector of length 2.
The order of x[i-1] and x[i-2] can also be swapped.

Usage

seq_rec2(
  inits = c(0L, 1L),
  n = 10L,
  s = c(0L, 0L),
  m = c(1L, 1L),
  inop = "+",
  form = 1L,
  rev = FALSE
)

Arguments

inits

a numeric (double or integer) vector of length 2, giving the initial values.
Any numbers are allowed, even negative and/or fractional numbers.
Note that numbers given must give valid results when passed to function f().

n

a single integer, giving the size of the numeric vector to generate.
NOTE: it must hold that n > 2.

s, m

numeric vectors of length 2 to be used in the formula.

inop

a single string, giving the arithmetic infix operator to be used.
Currently supported: "+", "-", "*", "/".
For a fibonacci sequence, inop = "+".

form

either 1 or 2, indicating which formula to be used (see Description section above).

rev

reverse the order of x[i-1] and x[-2].
For example, using form = 1:

If rev = FALSE (default), it holds:
x[i] = (s[1] + m[1] * x[i-1]) %inop% (s[2] + m[2] * x[i-2]).

If rev = TRUE, it holds:
x[i] = (s[1] + m[1] * x[i-2]) %inop% (s[2] + m[2] * x[i-1])

Value

A sequence of numbers.

Details

The default values of the arguments give the first 10 numbers of a regular Fibonacci sequence.
See examples for several number series created with this function.
This function is written in C++ using Rcpp for better performance.

Note

Do not supply NAs or NaNs to this function, as it cannot handle them.

Examples

seq_rec2() # by default gives Fibonacci numbers
#>  [1]  0  1  1  2  3  5  8 13 21 34
seq_rec2(inits = 2:1) # Lucas numbers
#>  [1]  2  1  3  4  7 11 18 29 47 76
c(1, seq_rec2(c(1, 2), inop = "*")) # Multiplicative Fibonacci
#>  [1]           1           1           2           2           4           8
#>  [7]          32         256        8192     2097152 17179869184
seq_rec2(m = c(2L, 1L)) # Pell numbers
#>  [1]   0   1   2   5  12  29  70 169 408 985
seq_rec2(inits = c(1, 0), m = c(0L, 2L)) # see https://oeis.org/A077957
#>  [1]  1  0  2  0  4  0  8  0 16  0
seq_rec2(m = c(1L, 2L)) # Jacobsthal numbers
#>  [1]   0   1   1   3   5  11  21  43  85 171