Formula with Keywords

Instead of the usual indexing types, formulas can be used to specify indices in arrays that can evaluate keywords.

The following keywords are available:

.M: the given margin/dimension; 0 if not relevant.

.Nms: the (dim)names at the given margin.

.N: the size of a given dimension (if .M is not 0) or else the length of x.

.I: equal to 1:.N.

.bi(...): a function to specify bilateral indices. See Details below.

.x: the input variable x itself.

These keywords can be used inside formulas to specify more advanced indices.

Here are a few examples of advanced indexing in arrays:

x <- matrix(1:20, 5, 4)
dimnames(x) <- list(month.abb[1:5], month.abb[1:4])
print(x)
#>     Jan Feb Mar Apr
#> Jan   1   6  11  16
#> Feb   2   7  12  17
#> Mar   3   8  13  18
#> Apr   4   9  14  19
#> May   5  10  15  20

# select first half of rows, select first & last column:
ss_x(x, n(~ 1:round(.N/2), ~ .bi(1, -1)))
#>     Jan Apr
#> Jan   1  16
#> Feb   2  17

# extract where rownames contain "r" & where colnames DO NOT contain "r":
ss_x(x, ~ grep("r", .Nms), use = c(1, -2))
#>     Jan Feb
#> Mar   3   8
#> Apr   4   9

# extract 1st row and 2nd column (i.e. where dim index equals margin):
ss_x(x, ~ .M)
#>     Feb
#> Jan   6

Details

Bilateral Indices
The bi(...) function is used to specify bilateral indices.
It's usage is similar to the c(...) function, in that multiple vectors can be concatenated to one.
In bi(), one can specify both positive and negative numbers.
Positive numbers specify indices work like normally.
Negative numbers specify indices from the end.
I.e. ~ bi(-2) would be equivalent to length(x) - 1 if x is a vector, or dim(x)[.M] - 1 if x is an array.

NOTE:
If one supplies a single ALTREP vector into bi(), i.e. bi(1:10), bi() will not materialize the ALTREP vector.
This keeps bi() highly efficient.