Mutable Atomic Classes

The mutable_atomic class is a mutable version of atomic classes.
It works exactly the same in all aspects as regular atomic classes, with only one real difference:
sb_set accepts mutable_atomic, but does not accept regular atomic.
See squarebrackets_PassByReference for details.

Like data.table, [<- performs R's default copy-on-modification semantics.
For modification by reference, use sb_set.

Exposed functions (besides the S3 methods):

is.mutable_atomic(): checks if an object is atomic.

as.mutable_atomic(): converts a regular atomic object to mutable_atomic.

couldb.mutable_atomic(): checks if an object could be mutable_atomic.
An objects can become mutable_atomic if it is one of the following types:
logical, integer, double, character, complex, raw.
bit64::integer64 type is also supported, since it is internally defined as double.

Usage

mutable_atomic(data, names = NULL, dim = NULL, dimnames = NULL)

as.mutable_atomic(x, ...)

is.mutable_atomic(x)

couldb.mutable_atomic(x)

# S3 method for mutable_atomic
[(x, ...)

# S3 method for mutable_atomic
[(x, ...) <- value

# S3 method for mutable_atomic
format(x, ...)

# S3 method for mutable_atomic
print(x, ...)

Arguments

data

atomic vector giving data to fill the mutable_atomic object.

names, dim, dimnames

see setNames and array.

x

an atomic object.

...

method dependent arguments.

value

see Extract.

Value

For as.mutable_atomic:

Converts an atomic object (vector, matrix, array) to the same object, but with additional class "mutable_atomic", and the additional attribute "typeof".

For is.mutable_atomic:

Returns TRUE if the object is atomic, has the class "mutable_atomic", has the correctly set attribute "typeof", and has an address that does not overlap with the addresses of base objects.

is.mutable_atomic returns FALSE otherwise.

For couldb.mutable_atomic:

Returns TRUE if the object is one of the following types:

logical, integer, double, character, complex, raw.

bit64::integer64 type is also supported, since it is internally defined as double.

Returns FALSE otherwise.

Warning

Always use mutable_atomic() or as.mutable_atomic to create a mutable object, as they make necessary checks.
Circumventing these checks may break things.

Examples

x <- mutable_atomic(
  1:20, dim = c(5, 4), dimnames = list(letters[1:5], letters[1:4])
)
x
#>   a  b  c  d
#> a 1  6 11 16
#> b 2  7 12 17
#> c 3  8 13 18
#> d 4  9 14 19
#> e 5 10 15 20
#> mutable_atomic 
#> typeof:  integer 

x <- matrix(1:10, ncol = 2)
x <- as.mutable_atomic(x)
is.mutable_atomic(x)
#> [1] TRUE
print(x)
#>      [,1] [,2]
#> [1,]    1    6
#> [2,]    2    7
#> [3,]    3    8
#> [4,]    4    9
#> [5,]    5   10
#> mutable_atomic 
#> typeof:  integer 
x[, 1]
#> [1] 1 2 3 4 5
#> mutable_atomic 
#> typeof:  integer 
x[] <- as.double(x) # notifies the user a copy is being made
#> copying on modification; for modification by reference, use `sb_set()`
print(x) # "typeof" attribute adjusted accordingly, and class still present
#>      [,1] [,2]
#> [1,]    1    6
#> [2,]    2    7
#> [3,]    3    8
#> [4,]    4    9
#> [5,]    5   10
#> mutable_atomic 
#> typeof:  double