Compute Grouped Indices

Given:

a sub-set function f;

the complete range of indices r of some object x;

and a grouping factor grp;

the idx_by() function takes indices f(r) per group grp.
The result of idx_by() can be supplied to the indexing arguments (see squarebrackets_indx_args) of:
sb_x, sb_rm, sb_mod, sb_set, or sb_coe,
to perform grouped subset operations.

Usage

idx_by(f, r, grp, parallel = FALSE, mc.cores = 1L)

Arguments

f

a subset function to be applied per group on r.
The function must accept a character or integer vector as input, and produce a character or integer vector as output.
For example, to subset the last element per group, specify:
f = last

r

an integer or character vector, giving the complete range of indices of an object.
For example: colnames(x), 1:nrow(x), etc.

grp

a factor giving the groups. Make sure its order corresponds to i and r, otherwise it makes no sense.

parallel, mc.cores

see BY.

Value

A vector of indices of the same type as r.

Examples

# vectors ====
(a <- 1:20)
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20
(grp <- factor(rep(letters[1:5], each = 4)))
#>  [1] a a a a b b b b c c c c d d d d e e e e
#> Levels: a b c d e

# get the last element of `a` for each group in `grp`:
i <- idx_by(last, seq_along(a), grp)
sb_x(cbind(a, grp), row = i)
#>       a grp
#> [1,]  4   1
#> [2,]  8   2
#> [3,] 12   3
#> [4,] 16   4
#> [5,] 20   5

# data.frame ====
x <- data.frame(
  a = sample(1:20),
  b = letters[1:20],
  group = factor(rep(letters[1:5], each = 4))
)
print(x)
#>     a b group
#> 1   1 a     a
#> 2   9 b     a
#> 3  18 c     a
#> 4  13 d     a
#> 5   6 e     b
#> 6  11 f     b
#> 7  15 g     b
#> 8   2 h     b
#> 9  20 i     c
#> 10  5 j     c
#> 11  7 k     c
#> 12 12 l     c
#> 13 14 m     d
#> 14 10 n     d
#> 15 19 o     d
#> 16  4 p     d
#> 17 17 q     e
#> 18 16 r     e
#> 19  3 s     e
#> 20  8 t     e
# get the first row for each group in data.frame `x`:
row <- idx_by(first, 1:nrow(x), x$group)
sb_x(x, row)
#>    a b group
#> 1  1 a     a
#> 2  6 e     b
#> 3 20 i     c
#> 4 14 m     d
#> 5 17 q     e
# get the first row for each group for which a > 10:
x2 <- sb_x(x, filter = ~ a > 10)
row <- na.omit(idx_by(first, 1:nrow(x2), x2$group))
sb_x(x2, row)
#>    a b group
#> 1 18 c     a
#> 2 11 f     b
#> 3 20 i     c
#> 4 14 m     d
#> 5 17 q     e