Unnest Tree-like List to Recursive 2d Array or Flattened Recursive Vector

[[, [[<-, sb2_rec, and sb2_reccom, can performing recursive subset operation on a nested list.
Such recursive subset operations only operate on a single element.
Performing recursive subset operations on multiple elements is not vectorized, and requires a (potentially slow) loop.

The lst_untree() funtion takes a nested tree-like list, and turns it into a 2d recursive array (i.e. a list-matrix), allowing vectorized subset operations to be performed on the list.
lst_untree() can also simply flatten the list, making it a non-nested list.
See the Examples section to understand how the list will be arranged and named.

The lst_nlists() counts the total number of recursive list-elements inside a list.

Usage

lst_nlists(x)

lst_untree(x, margin, use.names = FALSE)

Arguments

x

a tree-like nested list.

margin

a single integer, indicating how the result should be arranged:

margin = 0 produces a simple flattened recursive vector (i.e. list) without dimensions.

margin = 1 produces a 2D recursive array (i.e. a matrix of lists),
with length(x) rows and n columns,
where n = sapply(x, lst_nlists) |> max().
Empty elements will be filled with list(NULL).

margin = 2 pproduces a 2D recursive array (i.e. a matrix of lists),
with length(x) columns and n rows,
where n = sapply(x, lst_nlists) |> max().
Empty elements will be filled with list(NULL).

use.names

Boolean, indicating if the elements returned from lst_untree() should be named.
Names of nested elements, such as x[[c("A", "B", "C")]], will become "A.B.C", as that is the behaviour of the rapply function (which lst_untree() calls internally).
It is therefore advised not to use dots (".") in your list names, and use underscores ("_") instead, before calling lst_untree().
See the rrapply::rrapply function for renaming (and other forms of transforming) recursive subsets of lists.

Value

For lst_untree():

A non-nested (dimensional) list.

Note that if margin = 1 or margin = 2, lst_untree() returns a recursive matrix (i.e. a recursive array with 2 dimensions), not a data.frame.

(One advantage of a recursive matrix over a data.frame, is that a recursive matrix can have separate column names and regular names, whereas the names of a data.frame are necessarily equal to the column names).

For lst_nlists():

A single integer, giving the total number of recursive list-elements in the given list.

Examples

# show-casing how the list-elements are arranged and named ====

x <- list(
  A = list(
    A = list(A = "AAA", B = "AAB"),
    A = list(A  = "AA2A", B = "AA2B"),
    B = list(A = "ABA", B = "ABB"),
    C = letters
  ),
  Y = list(
    Z = list(Z = "YZZ", Y = "YZY"),
    Y = list(Z = "YYZ", Y = "YYY"),
    X = "YX"
  )
)

# simple flattened list:
y <- lst_untree(x, margin = 0, use.names = TRUE)
print(y)
#> $A.A.A
#> [1] "AAA"
#> 
#> $A.A.B
#> [1] "AAB"
#> 
#> $A.A.A
#> [1] "AA2A"
#> 
#> $A.A.B
#> [1] "AA2B"
#> 
#> $A.B.A
#> [1] "ABA"
#> 
#> $A.B.B
#> [1] "ABB"
#> 
#> $A.C
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m" "n" "o" "p" "q" "r" "s"
#> [20] "t" "u" "v" "w" "x" "y" "z"
#> 
#> $Y.Z.Z
#> [1] "YZZ"
#> 
#> $Y.Z.Y
#> [1] "YZY"
#> 
#> $Y.Y.Z
#> [1] "YYZ"
#> 
#> $Y.Y.Y
#> [1] "YYY"
#> 
#> $Y.X
#> [1] "YX"
#> 
y[["Y.Z.Y"]]
#> [1] "YZY"
x[[c("Y", "Z", "Y")]] # equivalent in the original list
#> [1] "YZY"

# un-tree row-wise:
sapply(x, lst_nlists) |> max() # number of columns `y` will have
#> [1] 7
y <- lst_untree(x, margin = 1, use.names = TRUE)
dim(y)
#> [1] 2 7
print(y)
#>   [,1]  [,2]  [,3]   [,4]   [,5]  [,6]  [,7]        
#> A "AAA" "AAB" "AA2A" "AA2B" "ABA" "ABB" character,26
#> Y "YZZ" "YZY" "YYZ"  "YYY"  "YX"  NULL  NULL        
#> attr(,"names")
#>  [1] "A.A.A" "Y.Z.Z" "A.A.B" "Y.Z.Y" "A.A.A" "Y.Y.Z" "A.A.B" "Y.Y.Y" "A.B.A"
#> [10] "Y.X"   "A.B.B" ""      "A.C"   ""     
y[["Y.Z.Y"]] # you can still use names for selecting/replacing
#> [1] "YZY"
y[1:2, 1:3] # vectorized selection of multiple recursive elements
#>   [,1]  [,2]  [,3]  
#> A "AAA" "AAB" "AA2A"
#> Y "YZZ" "YZY" "YYZ" 
sb2_x(y, n(1:2, 1:3), 1:2) # equivalent
#>   [,1]  [,2]  [,3]  
#> A "AAA" "AAB" "AA2A"
#> Y "YZZ" "YZY" "YYZ" 

# un-tree column-wise:
sapply(x, lst_nlists) |> max() # number of rows `y` will have
#> [1] 7
y <- lst_untree(x, margin = 2, use.names = TRUE)
dim(y)
#> [1] 7 2
print(y)
#>      A            Y    
#> [1,] "AAA"        "YZZ"
#> [2,] "AAB"        "YZY"
#> [3,] "AA2A"       "YYZ"
#> [4,] "AA2B"       "YYY"
#> [5,] "ABA"        "YX" 
#> [6,] "ABB"        NULL 
#> [7,] character,26 NULL 
#> attr(,"names")
#>  [1] "A.A.A" "A.A.B" "A.A.A" "A.A.B" "A.B.A" "A.B.B" "A.C"   "Y.Z.Z" "Y.Z.Y"
#> [10] "Y.Y.Z" "Y.Y.Y" "Y.X"   ""      ""     
y[["Y.Z.Y"]] # you can still use names for selecting/replacing
#> [1] "YZY"
y[1:3, 1:2] # vectorized selection of multiple recursive elements
#>      A      Y    
#> [1,] "AAA"  "YZZ"
#> [2,] "AAB"  "YZY"
#> [3,] "AA2A" "YYZ"
sb2_x(y, n(1:3, 1:2), 1:2) # equivalent
#>      A      Y    
#> [1,] "AAA"  "YZZ"
#> [2,] "AAB"  "YZY"
#> [3,] "AA2A" "YYZ"

################################################################################

# showcasing that only list-elements are recursively flattened ====
# i.e. atomic vectors in recursive subsets remain atomic

x <- lapply(1:10, \(x)list(sample(letters), sample(1:10)))

sapply(x, lst_nlists) |> max()
#> [1] 2
y <- lst_untree(x, margin = 1)
dim(y)
#> [1] 10  2
print(y)
#>       [,1]         [,2]      
#>  [1,] character,26 integer,10
#>  [2,] character,26 integer,10
#>  [3,] character,26 integer,10
#>  [4,] character,26 integer,10
#>  [5,] character,26 integer,10
#>  [6,] character,26 integer,10
#>  [7,] character,26 integer,10
#>  [8,] character,26 integer,10
#>  [9,] character,26 integer,10
#> [10,] character,26 integer,10

lst_untree(x, margin = 0)
#> [[1]]
#>  [1] "v" "i" "j" "b" "f" "g" "l" "n" "e" "u" "c" "t" "a" "r" "s" "z" "o" "x" "m"
#> [20] "w" "h" "p" "y" "k" "d" "q"
#> 
#> [[2]]
#>  [1]  2  8  1  3  9  7  6 10  4  5
#> 
#> [[3]]
#>  [1] "y" "u" "c" "m" "r" "t" "k" "d" "b" "i" "v" "n" "x" "l" "o" "a" "j" "g" "e"
#> [20] "s" "z" "w" "h" "f" "p" "q"
#> 
#> [[4]]
#>  [1]  7  8  6  1  4  9  3  2  5 10
#> 
#> [[5]]
#>  [1] "u" "j" "c" "d" "m" "f" "n" "b" "e" "r" "y" "l" "w" "x" "h" "t" "i" "a" "g"
#> [20] "o" "q" "k" "z" "s" "v" "p"
#> 
#> [[6]]
#>  [1]  9 10  2  1  3  7  6  4  5  8
#> 
#> [[7]]
#>  [1] "o" "w" "e" "t" "k" "z" "h" "g" "y" "l" "c" "r" "q" "u" "b" "s" "n" "a" "x"
#> [20] "j" "v" "i" "m" "p" "d" "f"
#> 
#> [[8]]
#>  [1]  7  4  3  1  9 10  5  6  2  8
#> 
#> [[9]]
#>  [1] "t" "v" "o" "r" "b" "j" "w" "e" "n" "y" "s" "p" "f" "c" "a" "h" "k" "z" "u"
#> [20] "i" "m" "q" "g" "d" "x" "l"
#> 
#> [[10]]
#>  [1]  8  7  2  5 10  9  4  3  6  1
#> 
#> [[11]]
#>  [1] "y" "m" "b" "w" "z" "s" "c" "q" "a" "i" "p" "j" "o" "k" "e" "x" "n" "h" "v"
#> [20] "d" "l" "g" "f" "t" "r" "u"
#> 
#> [[12]]
#>  [1]  4 10  3  6  5  7  9  8  2  1
#> 
#> [[13]]
#>  [1] "n" "t" "u" "f" "g" "v" "w" "e" "l" "h" "k" "p" "b" "j" "o" "q" "z" "y" "i"
#> [20] "a" "s" "r" "d" "m" "x" "c"
#> 
#> [[14]]
#>  [1]  1 10  9  4  2  3  6  5  7  8
#> 
#> [[15]]
#>  [1] "j" "i" "v" "n" "c" "l" "z" "t" "b" "s" "g" "p" "d" "u" "m" "o" "q" "r" "a"
#> [20] "y" "w" "e" "k" "f" "x" "h"
#> 
#> [[16]]
#>  [1]  1 10  9  4  8  6  2  3  7  5
#> 
#> [[17]]
#>  [1] "i" "l" "m" "n" "e" "h" "f" "a" "j" "c" "b" "g" "w" "r" "x" "p" "v" "t" "o"
#> [20] "k" "y" "q" "d" "z" "u" "s"
#> 
#> [[18]]
#>  [1] 10  7  3  5  1  6  9  8  2  4
#> 
#> [[19]]
#>  [1] "r" "s" "j" "m" "q" "n" "w" "k" "x" "e" "p" "v" "y" "g" "t" "h" "o" "f" "b"
#> [20] "d" "l" "z" "a" "i" "c" "u"
#> 
#> [[20]]
#>  [1]  9  4  3  6  7  2  8 10  5  1
#>